+++ internal/database/models.go
@@ -94,11 +94,6 @@ func getEngine() (*xorm.Engine, error) {
 			conf.Database.User, conf.Database.Password, host, port, conf.Database.Name, conf.Database.SSLMode, conf.Database.Schema)
 		driver = "pgx"
 
-	case "mssql":
-		conf.UseMSSQL = true
-		host, port := dbutil.ParseMSSQLHostPort(conf.Database.Host)
-		connStr = fmt.Sprintf("server=%s; port=%s; database=%s; user id=%s; password=%s;", host, port, conf.Database.Name, conf.Database.User, conf.Database.Password)
-
 	case "sqlite3":
 		if err := os.MkdirAll(path.Dir(conf.Database.Path), os.ModePerm); err != nil {
 			return nil, errors.Newf("create directories: %v", err)